Move redis store to a separate package
